requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
python-docx
sentence-transformers
openai
python-dotenv
requests
